+++ transforms/__testfixtures__/class-test2.input.js
