+++ transforms/__testfixtures__/export-default-class.output.js
@@ -5,16 +5,12 @@
 import React from 'React';
 
 export default class extends React.Component {
-  constructor(props, context) {
-    super(props, context);
-
-    this.state = {
-      foo: 'bar',
-    };
-  }
-
   static propTypes = {
     foo: React.PropTypes.string,
+  };
+
+  state = {
+    foo: 'bar',
   };
 
   render() {
